refactor: 청크 공통 모듈 분리, dedup 기준 수정 및 overlap 처리 개선
