+++ rag/scripts/check_qna_chunks.py
@@ -50,8 +50,6 @@ def main():
         chunk_id = str(row.get("chunk_id", ""))
         chunk_id_counter[chunk_id] += 1
         chunk_strategy_counter[str(row.get("chunk_strategy", ""))] += 1
-
-        chunk_id = str(row.get("chunk_id", ""))
         if "::" in chunk_id:
             # chunk_id format: {doc_id}::{source_stem}::{chunk_index}
             doc_id = chunk_id.rsplit("::", 1)[0]
