+++ rag/scripts/check_tl1_chunks.py
@@ -51,7 +51,11 @@ def main():
         chunk_id_counter[chunk_id] += 1
         chunk_strategy_counter[str(row.get("chunk_strategy", ""))] += 1
 
-        doc_id = str(row.get("doc_id", ""))
+        if "::" in chunk_id:
+            #chunk_id 형식 = {doc_id}::{source}::{chunk_index}
+            doc_id = chunk_id.rsplit("::", 1)[0]
+        else:
+            doc_id = str(row.get("doc_id", ""))
         chunk_index = int(row.get("chunk_index", -1))
         doc_to_indices[doc_id].append(chunk_index)
